move SingleColorValueAttribute to Manager
