+++ src/es6/canvas_widget/Manager.js
@@ -8,7 +8,6 @@ import "https://cdnjs.cloudflare.com/ajax/libs/jquery-contextmenu/2.9.2/jquery.c
 import "https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.13.2/jquery-ui.min.js";
 import "https://unpkg.com/jquery@3.6.0/dist/jquery.js";
 import interact from "interactjs";
-
 import { default as _ } from "lodash-es";
 import { Map as YMap, Text as YText } from "yjs";
 import { CONFIG } from "../config";
@@ -39,7 +38,6 @@ import { default as BooleanAttribute } from "./BooleanAttribute";
 import { default as FileAttribute } from "./FileAttribute";
 import KeySelectionValueSelectionValueListAttribute from "./KeySelectionValueSelectionValueListAttribute";
 import QuizAttribute from "./QuizAttribute";
-import { default as SingleColorValueAttribute } from "./SingleColorValueAttribute";
 import { default as SingleMultiLineValueAttribute } from "./SingleMultiLineValueAttribute";
 import { default as SingleSelectionAttribute } from "./SingleSelectionAttribute";
 import { default as SingleValueAttribute } from "./SingleValueAttribute";
@@ -51,6 +49,11 @@ import LogicalConjunctions from "./viewpoint/LogicalConjunctions";
 import LogicalOperator from "./viewpoint/LogicalOperator";
 import RenamingListAttribute from "./viewpoint/RenamingListAttribute";
 import ViewTypesUtil from "./viewpoint/ViewTypesUtil";
+
+const singleColorValueAttributeHtml = await loadHTML(
+  "../../templates/canvas_widget/single_value_attribute.html",
+  import.meta.url
+);
 
 const keySelectionValueSelectionValueAttributeHtml = await loadHTML(
   "../../templates/canvas_widget/key_value_attribute.html",
@@ -8358,6 +8361,81 @@ export class KeySelectionValueAttribute extends AbstractAttribute {
     };
 
     _$node.find(".key").append(_key.get$node());
+    _$node.find(".value").append(_value.get$node());
+  }
+}
+/**
+ * SingleColorValueAttribute
+ * @class canvas_widget.SingleColorValueAttribute
+ * @extends canvas_widget.AbstractAttribute
+ * @memberof canvas_widget
+ * @constructor
+ * @param {string} id Entity id
+ * @param {string} name Name of attribute
+ * @param {canvas_widget.AbstractEntity} subjectEntity Entity the attribute is assigned to
+ */
+export class SingleColorValueAttribute extends AbstractAttribute {
+  constructor(id, name, subjectEntity) {
+    super(id, name, subjectEntity);
+
+    /***
+     * Value object of value
+     * @type {canvas_widget.Value}
+     * @private
+     */
+    var _value = new Value(id, name, this, this.getRootSubjectEntity());
+
+    /**
+     * jQuery object of DOM node representing the node
+     * @type {jQuery}
+     * @private
+     */
+    var _$node = $(_.template(singleColorValueAttributeHtml)());
+
+    /**
+     * Set Value object of value
+     * @param {canvas_widget.Value} value
+     */
+    this.setValue = function (value) {
+      _value = value;
+    };
+
+    /**
+     * Get Value object of value
+     * @returns {canvas_widget.Value}
+     */
+    this.getValue = function () {
+      return _value;
+    };
+
+    /**
+     * jQuery object of DOM node representing the attribute
+     * @type {jQuery}
+     * @private
+     */
+    this.get$node = function () {
+      return _$node;
+    };
+
+    /**
+     * Get JSON representation of the attribute
+     * @returns {Object}
+     */
+    this.toJSON = function () {
+      var json = AbstractAttribute.prototype.toJSON.call(this);
+      json.value = _value.toJSON();
+      return json;
+    };
+
+    /**
+     * Set attribute value by its JSON representation
+     * @param json
+     */
+    this.setValueFromJSON = function (json) {
+      _value.setValueFromJSON(json.value);
+    };
+
+    _$node.find(".name").text(this.getName());
     _$node.find(".value").append(_value.get$node());
   }
 }
